patch various bits of libluatex to remove compiler warnings
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2306 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/image/writeimg.c
@@ -652,7 +652,7 @@ typedef image *img_entry;
 /* define img_ptr, img_array, & img_limit */
 /* avoid use of size_t */
 img_entry *img_ptr, *img_array = NULL;
-integer img_limit;
+size_t img_limit;
 
 
 integer img_to_array(image * img)
